% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{aliases}
\alias{aliases}
\alias{proc_benchmarking}
\alias{proc_tsraking}
\alias{macro_gseriestsbalancing}
\title{Function aliases}
\usage{
proc_benchmarking(...)

proc_tsraking(...)

macro_gseriestsbalancing(...)
}
\arguments{
\item{...}{Corresponding function arguments.}
}
\description{
\if{html,text}{(\emph{version française: 
\url{https://ferlmic.github.io/gstest/fr/reference/aliases.html}})}

\code{\link[=proc_benchmarking]{proc_benchmarking()}} is an alias for \code{\link[=benchmarking]{benchmarking()}}

\code{\link[=proc_tsraking]{proc_tsraking()}} is an alias for \code{\link[=tsraking]{tsraking()}}

\code{\link[=macro_gseriestsbalancing]{macro_gseriestsbalancing()}} is an alias for \code{\link[=tsbalancing]{tsbalancing()}}
}
\details{
The name of these aliases is reminiscent of the corresponding function's SAS\eqn{^\circledR}{®} origin
(PROC BENCHMARKING, PROC TSRAKING and macro \emph{\strong{GSeriesTSBalancing}} in G-Series 2.0). These aliases also
ensure backward compatibility with early development versions of the R package.

See the corresponding function for examples or the description of the arguments and the returned value:
\itemize{
\item \code{\link[=benchmarking]{benchmarking()}} for alias \code{\link[=proc_benchmarking]{proc_benchmarking()}}
\item \code{\link[=tsraking]{tsraking()}} for alias \code{\link[=proc_tsraking]{proc_tsraking()}}
\item \code{\link[=tsbalancing]{tsbalancing()}} for alias \code{\link[=macro_gseriestsbalancing]{macro_gseriestsbalancing()}}
}
}
\keyword{internal}
