% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstest-package.R
\docType{package}
\name{gstest-package}
\alias{gstest-package}
\title{Generalized Two-Sample Kolmogorov-Smirnov Test for Right-Censored Data}
\description{
An implementation of the generalized two-sample Kolmogorov-Smirnov test for Right-Censored Data; Fleming et al. (1980) \doi{10.2307/2556114}.
}
\author{
Kengo Nagashima
}
\keyword{package}
