% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstest.r
\name{gstest}
\alias{gstest}
\title{Generalized Two-Sample Kolmogorov-Smirnov Test for Right-Censored Data}
\usage{
gstest(formula, data)
}
\arguments{
\item{formula}{a formula object, which must have a \code{Surv} object as the
response on the left of the ~ operator and only one term on the right.
This test is a two-sample test, and this function only supports a two-level
covariate. This function treats 1 as an event and 0 as a censoring.}

\item{data}{a data frame in which to interpret the variables named in the
formula.}
}
\value{
\itemize{
\item \code{A}: Test statistic \eqn{A}; see Fleming et al. (1980).
\item \code{R}: Test statistic \eqn{R}; see Fleming et al. (1980).
\item \code{pval}: Two-sieded P-value.
}
}
\description{
This function can perform the generalized two-sample Kolmogorov-Smirnov test
for right-censored data that proposed by Fleming et al. (1980).
}
\examples{
library(survival)

# Example 1 of Fleming et al. (1980)
df1 <- data.frame(
  y = c(28,89,175,195,309,377,393,421,447,462,709,744,770,1106,1206,
        34,88,137,199,280,291,299,300,309,351,358,369,369,370,375,
        382,392,429,451,1119),
  event = c(1,1,1,1,1,0,0,0,0,1,0,0,0,0,0,
            1,1,1,1,1,1,0,0,1,1,1,1,1,1,1,1,1,0,1,0),
  arm = c(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
          1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1)
)
# A = 1.834, R = 0.653, Two-sided P = 0.00223
gstest(Surv(y, event) ~ arm, df1)
}
\references{
Fleming, D., O'Fallon, JR., O'Brien, P.C., and Harrington, DP. (1980).
Modified Kolmogorov-Smirnov test procedures with application to arbitrarily
right-censored data.
\emph{Biometrics}
\strong{36}(4): 607-625.
\url{https://doi.org/10.2307/2556114}
}
