% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseries-package.R
\docType{package}
\name{gstest-package}
\alias{gstest}
\alias{gstest-package}
\title{gstest: (EN) 'G-Series' in 'R' | (FR) 'G-Séries' en 'R'}
\description{
(EN) Statistics Canada's generalized system devoted to time series benchmarking and reconciliation. The methods used in 'G-Series' essentially come from Dagum and Cholette (2006) \doi{10.1007/0-387-35439-5}. | (FR) Système généralisé de Statistique Canada consacré à l'étalonnage et à la réconciliation de séries chronologiques. Les méthodes utilisées dans 'G-Séries' proviennent essentiellement de Dagum et Cholette (2006) \doi{10.1007/0-387-35439-5}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ferlmic.github.io/gstest/en/}
  \item \url{https://ferlmic.github.io/gstest/fr/}
  \item Report bugs at \url{https://github.com/ferlmic/gstest/issues/}
}

}
\author{
\strong{Maintainer}: Michel Ferland \email{michel.ferland@statcan.gc.ca}

Other contributors:
\itemize{
  \item Statistics Canada [copyright holder, funder]
}

}
\keyword{internal}
